+++ src/renderer/camera.cuh
@@ -6,11 +6,11 @@ class Camera
 {
 public:
     QUALIFIER_DEVICE Camera();
-    QUALIFIER_DEVICE Camera(const uint32_t spp, const uint64_t width, const uint64_t height,
+    QUALIFIER_DEVICE Camera(const uint32_t spp, const int width, const int height,
                             const float fov_x, const Vec3 eye, const Vec3 look_at, const Vec3 up);
 
-    QUALIFIER_DEVICE uint64_t width() const { return width_; }
-    QUALIFIER_DEVICE uint64_t height() const { return height_; }
+    QUALIFIER_DEVICE int width() const { return width_; }
+    QUALIFIER_DEVICE int height() const { return height_; }
     QUALIFIER_DEVICE uint32_t spp() const { return spp_; }
     QUALIFIER_DEVICE float spp_inv() const { return spp_inv_; }
     QUALIFIER_DEVICE float fov_x() const { return fov_x_; }
@@ -20,8 +20,8 @@ public:
     QUALIFIER_DEVICE Vec3 view_dy() const { return view_dy_; }
 
 private:
-    uint64_t width_;
-    uint64_t height_;
+    int width_;
+    int height_;
     uint32_t spp_;
     float spp_inv_;
     float fov_x_;
